+++ brtp/math/optim/discrete/__init__.py
@@ -1,0 +1,1 @@
+from .max_min_separation_selection import min_separation